split files into views and components
