+++ frontend/src/modules/Environments/views/EnvironmentView.js
@@ -37,10 +37,10 @@ import { SET_ERROR, useDispatch } from 'globalErrors';
 import { archiveEnvironment, getEnvironment, useClient } from 'services';
 import { KeyValueTagList, Stack, StackStatus } from 'modules/Shared';
 import EnvironmentNetworks from '../../Misc/views/Networks/NetworkList';
-import EnvironmentDatasets from './EnvironmentDatasets';
-import EnvironmentOverview from './EnvironmentOverview';
-import EnvironmentSubscriptions from './EnvironmentSubscriptions';
-import EnvironmentTeams from './EnvironmentTeams';
+import EnvironmentDatasets from '../components/EnvironmentDatasets';
+import EnvironmentOverview from '../components/EnvironmentOverview';
+import EnvironmentSubscriptions from '../components/EnvironmentSubscriptions';
+import EnvironmentTeams from '../components/EnvironmentTeams';
 
 const tabs = [
   { label: 'Overview', value: 'overview', icon: <Info fontSize="small" /> },
